single-prec compile flag done through tests, examples. Need to put in interfaces & makefile better

+++ src/finufft_c.h
@@ -3,13 +3,14 @@
 
 //#include <complex>
 #include <complex.h>
+#include "utils.h"
 
 #ifdef __cplusplus
   extern "C" {
 #endif
 
   //  int finufft1d1_c(int nj,double* xj,std::complex<double>* cj,int iflag, double eps,int ms, std::complex<double>* fk);
-  int finufft1d1_c(int nj,double* xj,double _Complex* cj,int iflag, double eps,int ms, double _Complex* fk);
+  int finufft1d1_c(int nj,FLT* xj,FLT _Complex* cj,int iflag, FLT eps,int ms, FLT _Complex* fk);
 
 #ifdef __cplusplus
 }
